refactor(api-server): split emergency routes into cohesive sub-modules
The emergency route file had grown to 1681 lines (51 KB) across six
concerns — protocols, contacts, incidents, broadcasts, drills, and
statistics — and showed up as a churn hotspot (1021 lines changed).

Split the single `routes/emergency.rs` into a directory module following
the established `routes/announcements/` convention:

- `emergency/shared.rs` — query/request DTOs, their `From` conversions,
  and the `is_emergency_manager` gate helper
- `emergency/{protocols,contacts,incidents,broadcasts,drills,statistics}.rs`
  — per-surface handlers, each exposing a `pub(super) fn router()`
- `emergency/mod.rs` — composes the sub-routers via `.merge()` and
  re-exports the public DTO types so `routes::emergency::*` resolves
  unchanged

Behavior-preserving: handler bodies, route paths, status codes, RLS
release semantics, and the public `router()` entry point nested at
`/api/v1/emergency` are all identical. No endpoint, auth, or response
change.
